Refactor date wrapper
